Fix reviews + add bool config
