+++ src/server/game/World/World.h
@@ -193,6 +193,7 @@ enum WorldBoolConfigs
     CONFIG_PREVENT_RENAME_CUSTOMIZATION,
     CONFIG_CREATURE_CHECK_INVALID_POSITION,
     CONFIG_GAME_OBJECT_CHECK_INVALID_POSITION,
+    CONFIG_LEGACY_BUFF_ENABLED,
     BOOL_CONFIG_VALUE_COUNT
 };
